+++ CHANGELOG.md
@@ -20,6 +20,10 @@
 
 ## Next v0.0.38(2023-04-XX)
 
+- feat: add [WinSet](https://www.autohotkey.com/docs/v1/lib/WinSet.htm) SubCommand
+  1. syntax-highlight
+  2. completion (default: `false`)
+  3. hover
 - test: add test from ahk-doc <https://github.com/AutoHotkey/AutoHotkeyDocs/blob/v1/docs/static/source/data_index.js>
 - fix: `hotkey, if` and `hotkey, key` syntax-highlight
   ![img](https://raw.githubusercontent.com/CoffeeChaton/vscode-autohotkey-NekoHelp/main/image/Changelog/v0-0-38-hotkey.png)